fix tests that were broken by adding backends package
